Simplify to Client only; deploy to github.io
Successfully deployed to https://swooby.github.io/openai-realtime-push-to-talk and running!

THIS IS "SAFE"!

The original example rightly shows how to have pseudo-production client request an ephemeral API key through a server, never exposing the dangerous API key to the client.
That is [almost] perfect for production code, but impractical for a **hosted anonymous demo** of how to use the API.
There is no way the **demo host** is going to use their own API key, even if safe/hidden, for anonymous users to freely use!
That would result in the owner of that hosted key getting billed!
No one is going to do that.
If someone wants to see a demo of how this API works **they need to provide THEIR OWN key**!
**The code is open source and proves it only ever sends the user's OpenAI api key directly to `api.openai.com`**.